Implementation in serizlizers for POC
Just a POC.
Haven't finished filters.
Haven't eliminate duplicate.
Haven't written tests.
